requirements.txt
fpdf
streamlit
pandas
Pillow
